--- src/main/java/roomescape/dto/PopularThemeRequestDto.java
@@ -1,9 +1,0 @@
-package roomescape.dto;
-
-import jakarta.validation.constraints.Max;
-
-public record PopularThemeRequestDto(
-        @Max(value = 15) int limit,
-        @Max(value = 10) int days
-) {
-}
